backend: rename oscompat back to dlhandle
