+++ gpt4all-backend/dlhandle.h
